Create unit test for class DeleteEmployerUseCaseAsync

+++ src/employers.application/UseCases/Employers/DeleteEmployerUseCaseAsync.cs
@@ -1,4 +1,5 @@
-﻿using employers.application.Interfaces.Empregado;
+﻿using employers.application.Exceptions.RegraNegocio;
+using employers.application.Interfaces.Empregado;
 using employers.domain.Interfaces.Repositories.Employers;
 using System.Threading.Tasks;
 
@@ -15,6 +16,9 @@ namespace employers.application.UseCases.Employers
 
         public async Task<int?> RunAsync(int id)
         {
+            if (id <= 0)
+                throw new RegranegocioException("Invalid ID!");
+
             var result = await _employerRepository.DeleteAsync(id);
             
             return result;
